requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pytesseract==0.3.10
